+++ packages/apps-engine/src/server/AppManager.ts
@@ -32,7 +32,7 @@ import { AppSignatureManager } from './managers/AppSignatureManager';
 import { UIActionButtonManager } from './managers/UIActionButtonManager';
 import type { IMarketplaceInfo } from './marketplace';
 import { defaultPermissions } from './permissions/AppPermissions';
-import type { DenoRuntimeSubprocessController } from './runtime/deno/AppsEngineDenoRuntime';
+import { EmptyRuntime } from './runtime/EmptyRuntime';
 import type { IAppStorageItem } from './storage';
 import { AppLogStorage, AppMetadataStorage } from './storage';
 import { AppSourceStorage } from './storage/AppSourceStorage';
@@ -58,6 +58,7 @@ export interface IAppManagerDeps {
 
 interface IPurgeAppConfigOpts {
 	keepScheduledJobs?: boolean;
+	keepSlashcommands?: boolean;
 }
 
 export class AppManager {
@@ -276,12 +277,7 @@ export class AppManager {
 				console.warn(`Error while compiling the App "${item.info.name} (${item.id})":`);
 				console.error(e);
 
-				const prl = new ProxiedApp(this, item, {
-					// Maybe we should have an "EmptyRuntime" class for this?
-					getStatus() {
-						return Promise.resolve(AppStatus.COMPILER_ERROR_DISABLED);
-					},
-				} as unknown as DenoRuntimeSubprocessController);
+				const prl = new ProxiedApp(this, item, new EmptyRuntime(item.id));
 
 				this.apps.set(item.id, prl);
 			}
@@ -359,7 +355,7 @@ export class AppManager {
 
 			this.listenerManager.releaseEssentialEvents(app);
 
-			app.getDenoRuntime().stopApp();
+			app.getRuntimeController().stopApp();
 		}
 
 		// Remove all the apps from the system now that we have unloaded everything
@@ -466,7 +462,7 @@ export class AppManager {
 			storageItem.status = await rl.getStatus();
 			// This is async, but we don't care since it only updates in the database
 			// and it should not mutate any properties we care about
-			await this.appMetadataStorage.update(storageItem).catch();
+			await this.appMetadataStorage.updateStatus(storageItem._id, storageItem.status).catch(() => {});
 		}
 
 		return isSetup;
@@ -487,7 +483,7 @@ export class AppManager {
 			await app.call(AppMethod.ONDISABLE).catch((e) => console.warn('Error while disabling:', e));
 		}
 
-		await this.purgeAppConfig(app, { keepScheduledJobs: true });
+		await this.purgeAppConfig(app, { keepScheduledJobs: true, keepSlashcommands: true });
 
 		await app.setStatus(status, silent);
 
@@ -499,7 +495,7 @@ export class AppManager {
 		storageItem.status = await app.getStatus();
 		// This is async, but we don't care since it only updates in the database
 		// and it should not mutate any properties we care about
-		await this.appMetadataStorage.update(storageItem).catch();
+		await this.appMetadataStorage.updateStatus(storageItem._id, storageItem.status).catch(() => {});
 
 		return true;
 	}
@@ -518,13 +514,13 @@ export class AppManager {
 		const storageItem = await this.appMetadataStorage.retrieveOne(id);
 
 		app.getStorageItem().marketplaceInfo = storageItem.marketplaceInfo;
-		await app.validateLicense().catch();
+		await app.validateLicense().catch(() => {});
 
 		storageItem.migrated = true;
 		storageItem.signature = await this.getSignatureManager().signApp(storageItem);
-		// This is async, but we don't care since it only updates in the database
-		// and it should not mutate any properties we care about
-		const stored = await this.appMetadataStorage.update(storageItem).catch();
+
+		const { marketplaceInfo, signature, migrated, _id } = storageItem;
+		const stored = await this.appMetadataStorage.updatePartialAndReturnDocument({ marketplaceInfo, signature, migrated, _id });
 
 		await this.updateLocal(stored, app);
 		await this.bridges
@@ -600,7 +596,7 @@ export class AppManager {
 
 		undoSteps.push(() =>
 			this.getRuntime()
-				.stopRuntime(app.getDenoRuntime())
+				.stopRuntime(app.getRuntimeController())
 				.catch(() => {}),
 		);
 
@@ -698,7 +694,7 @@ export class AppManager {
 
 		// Errors here don't really prevent the process from dying, so we don't really need to do anything on the catch
 		await this.getRuntime()
-			.stopRuntime(app.getDenoRuntime())
+			.stopRuntime(app.getRuntimeController())
 			.catch(() => {});
 
 		this.apps.delete(app.getID());
@@ -752,11 +748,13 @@ export class AppManager {
 		}
 
 		descriptor.signature = await this.signatureManager.signApp(descriptor);
-		const stored = await this.appMetadataStorage.update(descriptor);
+		const stored = await this.appMetadataStorage.updatePartialAndReturnDocument(descriptor, {
+			unsetPermissionsGranted: typeof permissionsGranted === 'undefined',
+		});
 
 		// Errors here don't really prevent the process from dying, so we don't really need to do anything on the catch
 		await this.getRuntime()
-			.stopRuntime(this.apps.get(old.id).getDenoRuntime())
+			.stopRuntime(this.apps.get(old.id).getRuntimeController())
 			.catch(() => {});
 
 		const app = await this.getCompiler().toSandBox(this, descriptor, result);
@@ -810,7 +808,7 @@ export class AppManager {
 
 				// Errors here don't really prevent the process from dying, so we don't really need to do anything on the catch
 				await this.getRuntime()
-					.stopRuntime(this.apps.get(stored.id).getDenoRuntime())
+					.stopRuntime(this.apps.get(stored.id).getRuntimeController())
 					.catch(() => {});
 
 				return this.getCompiler().toSandBox(this, stored, parseResult);
@@ -821,6 +819,7 @@ export class AppManager {
 			}
 		})();
 
+		// We don't keep slashcommands here as the update could potentially not provide the same list
 		await this.purgeAppConfig(app, { keepScheduledJobs: true });
 
 		this.apps.set(app.getID(), app);
@@ -906,7 +905,7 @@ export class AppManager {
 
 				appStorageItem.marketplaceInfo[0].subscriptionInfo = appInfo.subscriptionInfo;
 
-				return this.appMetadataStorage.update(appStorageItem);
+				return this.appMetadataStorage.updateMarketplaceInfo(appStorageItem._id, appStorageItem.marketplaceInfo);
 			}),
 		).catch();
 
@@ -942,7 +941,7 @@ export class AppManager {
 						const storageItem = app.getStorageItem();
 						storageItem.status = status;
 
-						return this.appMetadataStorage.update(storageItem).catch(console.error) as Promise<void>;
+						return this.appMetadataStorage.updateStatus(storageItem._id, storageItem.status).catch(console.error) as Promise<void>;
 					}),
 			),
 		);
@@ -1045,6 +1044,8 @@ export class AppManager {
 			await app.call(AppMethod.INITIALIZE);
 			await app.setStatus(AppStatus.INITIALIZED, silenceStatus);
 
+			await this.commandManager.registerCommands(app.getID());
+
 			result = true;
 		} catch (e) {
 			let status = AppStatus.ERROR_DISABLED;
@@ -1071,7 +1072,7 @@ export class AppManager {
 			// This is async, but we don't care since it only updates in the database
 			// and it should not mutate any properties we care about
 			storageItem.status = await app.getStatus();
-			await this.appMetadataStorage.update(storageItem).catch();
+			await this.appMetadataStorage.updateStatus(storageItem._id, storageItem.status).catch(() => {});
 		}
 
 		return result;
@@ -1081,9 +1082,13 @@ export class AppManager {
 		if (!opts.keepScheduledJobs) {
 			await this.schedulerManager.cleanUp(app.getID());
 		}
+
+		if (!opts.keepSlashcommands) {
+			await this.commandManager.unregisterCommands(app.getID());
+		}
+
 		this.listenerManager.unregisterListeners(app);
 		this.listenerManager.lockEssentialEvents(app);
-		await this.commandManager.unregisterCommands(app.getID());
 		this.externalComponentManager.unregisterExternalComponents(app.getID());
 		await this.apiManager.unregisterApis(app.getID());
 		this.accessorManager.purifyApp(app.getID());
@@ -1157,7 +1162,6 @@ export class AppManager {
 		}
 
 		if (enable) {
-			await this.commandManager.registerCommands(app.getID());
 			this.externalComponentManager.registerExternalComponents(app.getID());
 			await this.apiManager.registerApis(app.getID());
 			this.listenerManager.registerListeners(app);
@@ -1165,14 +1169,14 @@ export class AppManager {
 			this.videoConfProviderManager.registerProviders(app.getID());
 			await this.outboundCommunicationProviderManager.registerProviders(app.getID());
 		} else {
-			await this.purgeAppConfig(app);
+			await this.purgeAppConfig(app, { keepScheduledJobs: true, keepSlashcommands: true });
 		}
 
 		if (saveToDb) {
 			storageItem.status = status;
 			// This is async, but we don't care since it only updates in the database
 			// and it should not mutate any properties we care about
-			await this.appMetadataStorage.update(storageItem).catch();
+			await this.appMetadataStorage.updateStatus(storageItem._id, storageItem.status).catch(() => {});
 		}
 
 		await app.setStatus(status, silenceStatus);