+++ packages/ui-video-conf/src/VideoConfPopup/VideoConfPopupHeader.tsx
@@ -2,7 +2,7 @@ import { Box } from '@rocket.chat/fuselage';
 import type { ReactNode, ReactElement } from 'react';
 
 const VideoConfPopupHeader = ({ children }: { children: ReactNode }): ReactElement => (
-	<Box display='flex' justifyContent='space-between'>
+	<Box display='flex' minHeight='x28' justifyContent='space-between'>
 		{children}
 	</Box>
 );